aikalaillavalmisehkä

+++ src/api/resolvers/tagResolver.ts
@@ -1,4 +1,11 @@
-import {deleteTag, fetchAllTags, fetchTagsByMediaId} from '../models/tagModel';
+import {GraphQLError} from 'graphql';
+import {MyContext} from '../../local-types';
+import {
+  deleteTag,
+  fetchAllTags,
+  fetchTagsByMediaId,
+  postTag,
+} from '../models/tagModel';
 
 export default {
   MediaItem: {
@@ -12,8 +19,25 @@ export default {
     },
   },
   Mutation: {
-    deleteTag: async (_parent: undefined, args: {input: string}) => {
+    deleteTag: async (
+      _parent: undefined,
+      args: {input: string},
+      context: MyContext,
+    ) => {
+      if (!context.user || context.user.level_name !== 'Admin') {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'NOT_AUTHORIZED'},
+        });
+      }
       return await deleteTag(Number(args.input));
+    },
+    createTag: async (_parent: undefined, args: {tag_name: string}) => {
+      // capitalize first letter of tag_name because we want all tags to be the same
+      // format in the database so we can query them easily and check for duplicates
+      args.tag_name =
+        args.tag_name.charAt(0).toUpperCase() +
+        args.tag_name.slice(1).toLowerCase();
+      return await postTag(args.tag_name);
     },
   },
 };
